auth and add member to existing group by an already member

+++ laravel/app/Http/Controllers/LCController.php
@@ -29,11 +29,9 @@ class LCController extends Controller
      */
     public function store(Request $request)
     {
-        //
-        //if (Gate::denies('create-lc')) {
-        //    abort(403);
-        //}
-        //
+        if (Gate::denies('create-lc')) {
+            abort(403);
+        }
 
         $this->validate($request, [
             'city' => 'required|max:255',
@@ -69,11 +67,9 @@ class LCController extends Controller
     {
         $lc = LC::findOrFail($id);
 
-        //
-        //if (Gate::denies('edit-lc', $lc)) {
-        //    abort(403);
-        //}
-        //
+        if (Gate::denies('edit-lc', $lc)) {
+            abort(403);
+        }
 
         $this->validate($request, [
             'city' => 'required|max:255',
@@ -95,11 +91,9 @@ class LCController extends Controller
     {
         $lc = LC::findOrFail($id);
 
-        //
-        //if (Gate::denies('destroy-lc', $lc)) {
-        //    abort(403);
-        //}
-        //
+        if (Gate::denies('destroy-lc', $lc)) {
+            abort(403);
+        }
 
         $lc->delete();
         return response()->json("ok");
